fixed triangle counting (again)

+++ triangle.cpp
@@ -50,8 +50,8 @@ void read_stl(string fname, vector <triangle>&v) {
 
     char header_info[80] = "";
     char nTri[4];
-    unsigned long nTriLong;
-    unsigned long count = 0;
+    unsigned nTriLong;
+    unsigned count = 0;
 
     //read 80 byte header
     if (myFile) {
@@ -65,7 +65,7 @@ void read_stl(string fname, vector <triangle>&v) {
     //read 4-byte ulong
     if (myFile) {
         myFile.read (nTri, 4);
-        nTriLong = *((unsigned long*)nTri) ;
+        nTriLong = *((unsigned*)nTri) ;
         cout <<"Number of triangles: " << nTriLong << endl;
     }
     else{
